Move decoration style to the draggable view component
