+++ src/constants/props.ts
@@ -3,7 +3,7 @@ import type { SharedProps } from '../types';
 export const SHARED_PROPS: Required<SharedProps> = {
   activeItemOpacity: 1,
   activeItemScale: 1.1,
-  activeItemShadowOpacity: 0.15,
+  activeItemShadowOpacity: 0.2,
   dragEnabled: true,
   inactiveItemOpacity: 0.5,
   inactiveItemScale: 1
